refactor: extract payment callbacks into their own files + handle subscription user_audit_log
